Eliminate 'incompatible pointer type passed to GC_grow_table' gcc warning (fix of commit 642a4d1f2)
* allchblk.c (GC_dump_regions): Change type of i local variable from
unsigned to size_t.
* alloc.c [GC_ASSERTIONS] (GC_add_to_heap): Likewise.
* alloc.c [!NO_DEBUGGING] (GC_print_heap_sects): Likewise.
* blacklst.c (total_stack_black_listed): Likewise.
* checksums.c (GC_check_dirty): Likewise.
* alloc.c (GC_add_to_heap): Remove casts to size_t.
* extra/MacOS.c (perform_final_collection): Change type of
last_fo_entries local variable from word to size_t.
* include/private/gc_priv.h (GC_arrays._n_heap_sects,
GC_arrays._fo_entries): Change type from word to size_t.

+++ allchblk.c
@@ -187,7 +187,7 @@ STATIC int GC_hblk_fl_from_blocks(size_t blocks_needed)
 
   GC_API void GC_CALL GC_dump_regions(void)
   {
-    unsigned i;
+    size_t i;
 
     for (i = 0; i < GC_n_heap_sects; ++i) {
         ptr_t start = GC_heap_sects[i].hs_start;
